Fixed problem with nearby and issue with hot reload user

+++ frontend/lib/models/location.dart
@@ -6,8 +6,8 @@ class Location {
 
   Location.fromJson(Map<String, dynamic> parsedJson) {
     this.id = parsedJson['id'];
-    this.lat = parsedJson['lat'];
-    this.lng = parsedJson['lng'];
+    this.lat = parsedJson['lat'].toDouble();
+    this.lng = parsedJson['lng'].toDouble();
     this.show = parsedJson['show'];
   }
 }
